drm/i915/gvt: fix the racing condition when vgpu destroy
When the acrn-dm is killed, there could be possibility that the
vGPU is destroyed but there's still workloads running in worker_thread,
which leads to a kernel panic. This is because pick_next_workload didn't
check if the vGPU is being destroyed.
This patch fixed the issue mentioned above, by adding a check when
picking next workload.
Also, there's a new MPT interface vgpu_release() can help to release
unfinished workloads, so we use this interface to replace previous
vgpu_deactivate().

Tracked-On: projectacrn/acrn-hypervisor#3530
Signed-off-by: Min He <min.he@intel.com>
Reviewed-by: Zhao Yakui <yakui.zhao@intel.com>